+++ CdIts.Caffoa.Cli/Model/AdditionalInterfaceModel.cs
@@ -1,0 +1,8 @@
+namespace CdIts.Caffoa.Cli.Model;
+
+public class AdditionalInterfaceModel
+{
+    public string VariableName { get; set; } = "";
+    public string ParameterType { get; set; } = "";
+    public string Initializer { get; set; } = "";
+}