+++ demo/DemoV3/Services/DemoV3Service.cs
@@ -17,7 +17,7 @@ namespace DemoV3.Services
     {
         private readonly UserRepository<UserWithId> _users = new UserRepository<UserWithId>();
         private readonly UserRepository<GuestUser> _guests = new UserRepository<GuestUser>();
-        private IContractResolver _responseContractResolver = new RemoveRequiredContractResolver();
+        private readonly IContractResolver _responseContractResolver = new RemoveRequiredContractResolver();
         public IDemoV3Service Instance(HttpRequest request)
         {
             return this;
@@ -35,18 +35,27 @@ namespace DemoV3.Services
             }
         }
 
-        public async Task<IEnumerable<AnyCompleteUser>> UsersGetAsync()
+        public async Task<IEnumerable<AnyCompleteUser>> UsersGetAsync(int offset = 0, int limit = 1000)
         {
             var result = new List<AnyCompleteUser>();
             result.AddRange(await _users.List());
             result.AddRange(await _guests.List());
-            return result;
+            return result.Skip(offset).Take(limit);
         }
 
-        public async Task<IEnumerable<User>> UsersGetByBirthdateAsync(DateTime date)
+        public async Task<IEnumerable<User>> UsersGetByBirthdateAsync(DateOnly date)
         {
             var users = await _users.List();
             return users.Where(u => u.Birthdate >= date);
+        }
+
+        public async Task<IEnumerable<User>> UsersSearchByDateAsync(DateOnly before, DateOnly after, int? maxResults = null)
+        {
+            var users = await _users.List();
+            var results = users.Where(u => u.Birthdate < before && u.Birthdate > after);
+            if (maxResults is > 0)
+                results = results.Take(maxResults.Value);
+            return results;
         }
 
         public async Task<AnyCompleteUser> UserPostAsync(User payload)