+++ changelog.md
@@ -1,5 +1,11 @@
 # caffoa changelog
 
+### 1.1.0
+* new option: use `parsePathParameters` to let caffoa handle the path parameter parsing instead of azure functions.
+Additionally, you can individualize the parsing, by implementing your own logic.
+* new option: use `parseQueryParameters` to parse query parameters and add them to the interface parameters.
+Additionally, required path options then raise an error via the ErrorHandler interface when empty.
+
 ### 1.0.4
 * Handle special chars in properties, classes and operations